fix issues with crc checksum calculation

+++ self-build.sh
@@ -3,13 +3,17 @@
 BASEDIR=$(pwd)/$(dirname $0)
 cd $BASEDIR
 
-rm build/self.sh
+if [ -e build/self.sh ]
+then
+    rm build/self.sh
+fi
 
 if [ ! -e dbs ]
 then
     cp build/main dbs
 fi
 
+mkdir -p build
 ./dbs src/main.dbs | tee build/self.sh
 chmod 755 build/self.sh
 
@@ -17,4 +21,4 @@ echo
 echo running ...
 build/self.sh || exit 1
 
-#mv build/main dbs
+mv build/main dbs
